+++ requirements.txt
@@ -2,7 +2,7 @@ flask==2.3.3
 flask-cors==4.0.0
 flask-sqlalchemy==3.0.5
 transformers==4.35.2
-torch==2.1.0
+torch==2.2.0
 scikit-learn==1.3.0
 pandas==2.0.3
 numpy==1.24.3
@@ -14,4 +14,5 @@ pytest-cov==4.1.0
 flake8==6.0.0
 bandit==1.7.5
 safety==2.3.4
+pyspark==3.5.0
